move add mime type guesser pass to the component
