+++ src/Symfony/Component/Mime/composer.json
@@ -18,6 +18,9 @@
     "require": {
         "php": "^7.1.3"
     },
+    "require-dev": {
+        "symfony/dependency-injection": "~3.4|^4.1"
+    },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Mime\\": "" },
         "exclude-from-classmap": [
